+++ plugin/action/join_template/join_template.go
@@ -1,9 +1,6 @@
 package join_template
 
 import (
-	"strings"
-	"unicode"
-
 	"github.com/ozontech/file.d/cfg"
 	"github.com/ozontech/file.d/fd"
 	"github.com/ozontech/file.d/logger"
@@ -131,209 +128,4 @@ func (p *Plugin) Stop() {
 
 func (p *Plugin) Do(event *pipeline.Event) pipeline.ActionResult {
 	return p.jp.Do(event)
-}
-
-func goPanicStartCheck(s string) bool {
-	return strings.HasPrefix(s, "panic:") ||
-		strings.Contains(s, "http: panic serving") ||
-		strings.HasPrefix(s, "fatal error:")
-}
-
-func checkOnlySpaces(s string) bool {
-	for _, c := range s {
-		if !unicode.IsSpace(c) {
-			return false
-		}
-	}
-
-	return true
-}
-
-const (
-	checkGoroutineIDPrefix = "goroutine "
-	checkGoroutineIDSuffix = " ["
-)
-
-// may be error: not only first occurrence counts
-func checkGoroutineID(s string) bool {
-	i := strings.Index(s, checkGoroutineIDPrefix)
-	if i == -1 {
-		return false
-	}
-
-	s = s[i+len(checkGoroutineIDPrefix):]
-
-	i = strings.Index(s, checkGoroutineIDSuffix)
-	if i == -1 {
-		return false
-	}
-
-	// no goroutine id found
-	if i == 0 {
-		return false
-	}
-
-	s = s[:i]
-	return checkOnlyDigits(s)
-}
-
-func checkOnlyDigits(s string) bool {
-	for _, c := range s {
-		if !unicode.IsDigit(c) {
-			return false
-		}
-	}
-
-	return true
-}
-
-func goPanicContinueCheck(s string) bool {
-	return checkOnlySpaces(s) ||
-		checkGoroutineID(s) ||
-		checkLineNumberAndFile(s) ||
-		checkCreatedBy(s) ||
-		strings.HasPrefix(s, "[signal") ||
-		checkPanicAddress(s) ||
-		strings.Contains(s, "panic:") ||
-		checkMethodCall(s)
-}
-
-const lineSuffix = ".go:"
-
-// may be error: not only first occurrence counts
-func checkLineNumberAndFile(s string) bool {
-	i := strings.Index(s, lineSuffix)
-	if i == -1 {
-		return false
-	}
-
-	s = s[i+len(lineSuffix):]
-
-	i = 0
-	for ; i < len(s) && unicode.IsDigit(rune(s[i])); i++ {
-	}
-
-	return i > 0
-}
-
-const checkCreatedBySubstr = "created by "
-
-func checkCreatedBy(s string) bool {
-	i := strings.Index(s, checkCreatedBySubstr)
-	if i == -1 {
-		return false
-	}
-
-	s = s[i+len(checkCreatedBySubstr):]
-
-	return strings.IndexByte(s, '.') != -1
-}
-
-func isIDChar(c byte) bool {
-	isLowerCase := 'a' <= c && c <= 'z'
-	isUpperCase := 'A' <= c && c <= 'Z'
-	isDigit := '0' <= c && c <= '9'
-	return isLowerCase || isUpperCase || isDigit || c == '_'
-}
-
-func isIDCharNoDigit(c byte) bool {
-	isLowerCase := 'a' <= c && c <= 'z'
-	isUpperCase := 'A' <= c && c <= 'Z'
-	return isLowerCase || isUpperCase || c == '_'
-}
-
-func checkMethodCallIndex(s string, pos int) bool {
-	if s[pos] != ')' {
-		return false
-	}
-
-	i := strings.LastIndex(s[:pos], "(")
-	if i == -1 {
-		return false
-	}
-
-	right := i - 1
-	left := right
-	for ; left >= 0 && isIDChar(s[left]); left-- {
-	}
-
-	if left == right {
-		return false
-	}
-
-	if left == -1 {
-		return false
-	}
-
-	if s[left] != '.' {
-		return false
-	}
-
-	left--
-	if left >= 0 && s[left] == ')' {
-		left--
-	}
-
-	return checkEndsWithClassName(s[:left])
-}
-
-func checkEndsWithClassName(s string) bool {
-	i := len(s) - 1
-	for ; i >= 0; i-- {
-		c := s[i]
-		if isIDCharNoDigit(c) {
-			break
-		}
-
-		isDigit := '0' <= c && c <= '9'
-		if !isDigit {
-			return false
-		}
-	}
-
-	return i >= 0
-}
-
-func checkMethodCall(s string) bool {
-	for i := len(s) - 1; i >= 0; i-- {
-		res := checkMethodCallIndex(s, i)
-		if res {
-			return true
-		}
-	}
-
-	return false
-}
-
-const (
-	checkPanicPrefix1 = "panic"
-	checkPanicPrefix2 = "0x"
-)
-
-func checkPanicAddress(s string) bool {
-	i := strings.Index(s, checkPanicPrefix1)
-	if i == -1 {
-		return false
-	}
-
-	s = s[i+len(checkPanicPrefix1):]
-
-	i = strings.Index(s, checkPanicPrefix2)
-	if i == -1 {
-		return false
-	}
-
-	// no chars between parts
-	if i == 0 {
-		return false
-	}
-
-	s = s[i+len(checkPanicPrefix2):]
-
-	if s == "" {
-		return false
-	}
-
-	c := s[0]
-	return '0' <= c && c <= '9' || 'a' <= c && c <= 'f'
 }
